refactor: 패키지 간소화
