+++ src/main/java/com/codingmate/controller/api/AnswerController.java
@@ -1,4 +1,4 @@
-package com.codingmate.controller.api.answer;
+package com.codingmate.controller.api;
 
 import com.codingmate.common.response.ResponseDto;
 import com.codingmate.common.response.ResponseMessage;
