+++ src/main/java/com/codingmate/controller/ui/AnswerUIController.java
@@ -1,4 +1,4 @@
-package com.codingmate.controller.ui.answer;
+package com.codingmate.controller.ui;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
