+++ src/main/java/com/codingmate/controller/ui/AuthUIController.java
@@ -1,13 +1,18 @@
-package com.codingmate.controller.ui.auth;
+package com.codingmate.controller.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/login")
-public class LoginController {
-    @RequestMapping
+public class AuthUIController {
+    @RequestMapping("/login")
     public String login() {
         return "auth/login";
+    }
+
+    @RequestMapping("/register")
+    public String register() {
+        return "auth/register";
     }
 }
