--- src/main/java/com/codingmate/controller/ui/auth/RegisterController.java
@@ -1,14 +1,0 @@
-package com.codingmate.controller.ui.auth;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/register")
-public class RegisterController {
-
-    @RequestMapping
-    public String register() {
-        return "auth/register";
-    }
-}
